BUG: El càlcul de l'espai d'un DVD es feia malament, ja que la capacitat en bytes d'un DVD no es pot expressar en unsigned long int. S'ha passat d'unsigned long int a double. Ticket: #326. Tambe queda solucionat la resta de problemes del ticket, ja que esdevenia un overflow al unsigned long al calcular les diferents capacitats.

+++ starviewer/src/inputoutput/cacheimagedal.h
@@ -51,7 +51,7 @@ public:
      * @param size Mida de les imatges que compleixen la màscara en bytes
      * @return retorna estat del mètode
      */
-    Status imageSize( DicomMask mask , unsigned long &size );
+    Status imageSize( DicomMask mask , double &size );
 
 
     /** Esborra les imatges que tinguin el studyUID passat per paràmetre
